backend/pyproject.toml
[project]
name = "tsmc-backend"
version = "0.1.0"
description = "FastAPI backend for TSMC"
readme = "../README.md"
requires-python = ">=3.12"
dependencies = [
  "aiosqlite>=0.21.0",
  "fastapi>=0.115.0",
  "greenlet>=3.2.4",
  "httpx>=0.28.1",
  "orjson>=3.10.0",
  "pydantic-settings>=2.10.1",
  "sqlalchemy>=2.0.44",
  "uvicorn[standard]>=0.35.0",
  "uvloop>=0.22.1; platform_system != 'Windows'",
]

[dependency-groups]
dev = [
  "pytest>=8.4.2",
  "pytest-asyncio>=1.2.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[tool.uv]
package = false
